feat: make LinkedList<T> File
