+++ Calculator/Calculator/ViewController.swift
@@ -14,46 +14,6 @@ class ViewController: UIViewController {
     }
 }
 
-class LinkedList<T> {
-    var head: Node<T>?
-    
-    var count: Int {
-        guard var node = head else {
-            return 0
-        }
-        var count = 1
-        while let next = node.next {
-            node = next
-            count += 1
-        }
-        return count
-    }
-    
-    func append(data: T?) {
-        if head == nil {
-            head = Node(data: data)
-            return
-        }
-        var node = head
-        while node?.next != nil {
-            node = node?.next
-        }
-        node?.next = Node(data: data)
-    }
-    
-    func remove() {
-        if head == nil {
-            return
-        }
-        head = head?.next
-    }
-    func removeAll() {
-        while head != nil {
-            head = head?.next
-        }
-    }
-}
-
 class Node<T> {
     var data: T?
     var next: Node?
